Adhere to common pattern even in last section of src/cmd/cas.x
All previous sections end with the sequence x/mn/r, but the section
:c9 ends with only x/mn. This is likely not correct.

Running

cd src/cmd
../../tools/as7 -o cas cas.s
../../tools/a7out cas cas2out cas.x

with and without this commit reveals a difference in cas2out.
Besides having the last number not be truncated, the trailing
statement for the :c9 entry:

c9=x-b+0264000

will only be emitted if the section ends with x/mn/r. This makes
cas emit well-formed output.