Fix out of range octal constant in Space Travel.
The scans of Space Travel, specifically 12-92-119.pdf page 26,
indicates that the literal stored at the location of the crash
variable should be -028. This appears to be out of range for
an octal number, and indeed the assembler complains about it.

It is unknown whether this is an issue with the assembler, but
it appears unlikely. I opted to change this to the legal value
-027. Changing the value does make Space Travel compile
successfully.